spelling, setup changes
Fixed typos in inclusions.py (prevented trails from showing up), and in
models.py (printing a Trail caused an exception). Finished working with
the setup procedure (code to handle only new additions is not set up,
neither is code to run the procedures for editing once all the data is
in the database, but ignoring those things, the full database can be set
up as of now). Also, cleaned up extra files in Inclusions module that
weren't used.