Inclusions, views, Trails, markers
Moved inclusions to a new folder, and added a new custom filter to
append a number to a string (and used that in places where add was not
working). Updated some models to fixed up their __unicode__ functions.
Fixed a typo in Layout.html. Fixed up much of Trails.html, which is now
mostly working. Added a new URL and view function for requesting markers
for the Trail map, function is mostly done.