Fixed up a few errors in passing along inital data
ASP tags where still in place in MapView.html to pass along
initial data to the JS functions. These have been replaced
with the appropriate Django tags. Fixed a path error in
the main JS script that had Google Maps API looking in the
wrong directory for the panorama files.

The panoramas are now displayed on the site. But, the functions
to then load the rest of the panoramas are not yet in place, and
links are screwed up. But, things are being displayed which is good.