Added PoI pages.
Added pages for all points of interest as well as categories. Made some
other changes along the way. Made links in menu bar use url tag (were
hard coded). Did the same for links to trails in Trails.html were
hard coded as part of a with tag. Moved .ogg for audio above .mp3 since
it is more widely supported. Changed how the database stores the audio
info for points of interest. They no longer include extensions, so
it must be manually appended in all <source> tags to make it easier
to handle multiple file types (the same was not done for photos).