Completed setup stuff.
Completed the function to fix links. The file fix_links.txt defines the
format for handling it. Also added some descriptive comments for what
the parameters for the data insertion function are. Changed settings.py
to use a small file to hold USER and PASSWORD for database which wasn't
needed for development (and the values used were just trash anyway) but
now that it's done that has to change to something more sustainable and
secure. Updated Readme with real information.